requirements.txt
Flask
torch
torchvision
Pillow
numpy
requests
gunicorn
